pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-wtf>=1.2.2",
    "injector>=0.22.0",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
]
